backend/requirements.txt
# FastAPI and server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9

# Data validation and serialization
pydantic>=2.6.0
pydantic-settings>=2.1.0

# Database
sqlalchemy>=2.0.0
asyncpg>=0.29.0
alembic>=1.13.0

# Caching
redis>=5.0.0

# Distributed Task Queue
arq>=0.26.0

# Authentication
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0

# Cryptography
cryptography>=42.0.0

# Trading SDKs
hyperliquid-python-sdk>=0.8.0
ccxt>=4.2.0
eth-account>=0.11.0

# AI SDKs
anthropic>=0.40.0
openai>=1.0.0
google-generativeai>=0.8.0

# HTTP client
httpx>=0.26.0
aiohttp>=3.9.0

# Utilities
python-dotenv>=1.0.0
tenacity>=8.2.0
pybreaker>=1.2.0

# Monitoring
prometheus-client>=0.19.0
sentry-sdk[fastapi]>=2.0.0

# Email
email-validator>=2.1.0
resend>=2.0.0

# Development
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0
aiosqlite>=0.19.0  # For async SQLite in tests
testcontainers>=4.0.0  # For integration tests with real databases
fakeredis>=2.20.0  # For Redis mocking
black>=24.1.0
ruff>=0.1.0
